requirements.txt
fastapi
uvicorn
pandas
numpy
Pillow
python-multipart
scikit-image